+++ releases.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
   - `Thread::Queue` transfers ownership of objects popped from it.
+  - Add support for `logger:` option in `Async::Safe.enable!` which logs violations instead of raising errors.
 
 ## v0.1.0